requirements.txt
streamlit
jsonschema>=4.21.0
pytest>=8.2.0
pyyaml>=6.0.0
pdfplumber>=0.11.0
pytesseract>=0.3.10
pillow>=10.4.0
